requirements.txt
streamlit
python-dotenv
pyodbc